Modularize Home Page (#575)
* Pull Twitch feature into own component file

* Pull events feature into own file

* Move academia callout into own section

* Move community share your stuff into new file

* Move hero content into own component
