+++ src/components/pages/home/index.js
@@ -1,0 +1,13 @@
+import AcademiaFeature from './academia-feature';
+import CommunityFeature from './community-feature';
+import EventsFeature from './events-feature';
+import Hero from './hero';
+import TwitchFeature from './twitch-feature';
+
+export {
+    AcademiaFeature,
+    CommunityFeature,
+    EventsFeature,
+    Hero,
+    TwitchFeature,
+};
